requirements.txt
fastai==2.7.12
librosa==0.9.2
matplotlib==3.7.1
numpy==1.22.2
opencv_python==4.7.0.72
Pillow==10.2.0
psola==0.0.1
PyYAML==6.0.1
soundfile==0.12.1
torch==2.2.0
torchaudio==2.2.0
torchvision==0.15.1
tqdm==4.65.0